cam-v1/camera_backend/auth-api/pyproject.toml
[tool.poetry]
name = "auth-api"
version = "0.1.0"
description = ""
authors = ["KiyoakiYamamoto <kiyo4447aki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
gunicorn = "^23.0.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pymysql = "^1.1.1"
pydantic = "^2.10.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptograpy = "^0.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

package-mode = false